+++ spec/units/gandi_v5_spec.rb
@@ -15,6 +15,12 @@ describe GandiV5 do
     expect(described_class.domains(param: :value)).to be returns
   end
 
+  it '.mailboxes' do
+    returns = double Array
+    expect(GandiV5::Email::Mailbox).to receive(:list).with('example.com', param: :value).and_return(returns)
+    expect(described_class.mailboxes('example.com', param: :value)).to be returns
+  end
+
   it 'Has a version' do
     expect(defined?(GandiV5::VERSION)).to eq 'constant'
     expect(GandiV5::VERSION).to_not be_empty